handwritten_digit_recognition/requirements.txt
torch>=1.8.0
torchvision>=0.9.0
numpy>=1.21.0
matplotlib>=3.4.0
pillow>=8.3.0
